requirements.txt
streamlit==1.33.0
openai
pillow